using size_t output in printf [See #375]

+++ trunk/lib/vlm-c/shmem.c
@@ -1,5 +1,5 @@
 /**
- * $Id: shmem.c,v 1.20 2009-09-06 18:44:11 ylafon Exp $
+ * $Id: shmem.c,v 1.22 2010-08-09 12:40:04 ylafon Exp $
  *
  * (c) 2008 by Yves Lafon
  *      See COPYING file for copying and redistribution conditions.
@@ -252,9 +252,9 @@ int copy_grib_array_to_shmem(int shmid, winds_prev *windtable, void *memseg) {
   /* be sure to lock the semaphore before using this function */
   nb_bytes = nb_prevs * sizeof(winds);
   printf("Got %d entries in windtable\n", windtable->nb_prevs);
-  printf("Bytes used: %ld (winds struct size is %ld)\n",nb_bytes,sizeof(winds));
+  printf("Bytes used: %ld (winds struct size is %zu)\n",nb_bytes,sizeof(winds));
   ok = (shminfo.shm_segsz > nb_bytes);
-  printf("Segment size: %ld %s\n",  (long)shminfo.shm_segsz, 
+  printf("Segment size: %zu %s\n",  shminfo.shm_segsz, 
 	 (ok) ? "OK" : "NOT OK");
 
   if (!ok) {
@@ -424,7 +424,7 @@ int copy_polar_array_to_shmem(int shmid, boat_polar_list *polars,
 
   printf("Bytes used: %ld\n", needed_bytes);
   ok = (shminfo.shm_segsz >= needed_bytes);
-  printf("Segment size: %ld %s\n",  (long)shminfo.shm_segsz, 
+  printf("Segment size: %zu %s\n",  shminfo.shm_segsz, 
 	 (ok) ? "OK" : "NOT OK");
   
   if (!ok) {
